backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
aiomysql
alembic
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
redis
httpx
twilio
geopy
numpy
scikit-learn
python-dateutil